小壳快读全本电子书 合并

+++ common_book/src/kdqbdzs/java/com/intelligent/reader/adapter/SearchSuggestAdapter.java
@@ -2,7 +2,6 @@ package com.intelligent.reader.adapter;
 
 import com.intelligent.reader.R;
 
-import net.lzbook.kit.utils.AppLog;
 import net.lzbook.kit.utils.AppUtils;
 
 import android.content.Context;
@@ -15,13 +14,9 @@ import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import net.lzbook.kit.constants.Constants;
 import net.lzbook.kit.data.search.SearchCommonBean;
-import net.lzbook.kit.utils.AppUtils;
 
 import java.util.List;
-
-import iyouqu.theme.ThemeHelper;
 
 /**
  * Created by Administrator on 2016/12/16 0016.
@@ -30,15 +25,12 @@ public class SearchSuggestAdapter extends BaseAdapter {
 
     private Context mContext;
     private List<SearchCommonBean> mData;
-    private String editInput;
-    private ThemeHelper mThemeHelper;
+    private String editInput = "";
 
     public SearchSuggestAdapter(Context context, List<SearchCommonBean> mData, String editInput) {
         this.mContext = context;
         this.mData = mData;
         this.editInput = editInput;
-        mThemeHelper = new ThemeHelper(mContext);
-
     }
 
     @Override
@@ -62,7 +54,7 @@ public class SearchSuggestAdapter extends BaseAdapter {
         if (convertView == null) {
             try {
                 LayoutInflater inflater = LayoutInflater.from(mContext);
-                convertView = inflater.inflate(R.layout.lv_searchbook_item, parent, false);
+                convertView = inflater.inflate(R.layout.item_search_suggest, parent, false);
             } catch (InflateException e) {
                 e.printStackTrace();
             }
@@ -88,16 +80,13 @@ public class SearchSuggestAdapter extends BaseAdapter {
             hodler.iv_type.setImageResource(R.drawable.search_transparent);
         }
         String content = bean.getSuggest();
+        String finalInput = "";
 
-        String finalInput = AppUtils.deleteAllIllegalChar(editInput);
-        if (mThemeHelper != null) {
-            if (mThemeHelper.isNight()) {
-                content = content.replaceAll(finalInput, "<font color='#90826b'>" + finalInput + "</font>");
-            } else {
-                content = content.replaceAll(finalInput, "<font color='#ff6600'>" + finalInput + "</font>");
-            }
+        if (editInput != null) {
+            finalInput = AppUtils.deleteAllIllegalChar(editInput);
         }
 
+        content = content.replaceAll(finalInput, "<font color='#ff6600'>" + finalInput + "</font>");
 
         hodler.tv_2.setText(Html.fromHtml(content));
         return convertView;
@@ -109,13 +98,13 @@ public class SearchSuggestAdapter extends BaseAdapter {
         }
     }
 
+    public void setEditInput(String editInput) {
+        this.editInput = editInput;
+    }
+
     private static class ViewHolder {
         TextView tv_2;
         ImageView iv_type;
-    }
-
-    public void setEditInput(String editInput) {
-        this.editInput = editInput;
     }
 
 }
